aula 03

+++ app/views/tarefa_views.py
@@ -3,11 +3,13 @@ from django.shortcuts import render, redirect
 from ..forms import TarefaForm
 from ..entidades.tarefas import Tarefa
 from ..services import tarefa_service
+from django.http import HttpResponse
+
 
 
 @login_required()
 def listar_tarefas(request):
-    tarefas = tarefa_service.listar_tarefas()
+    tarefas = tarefa_service.listar_tarefas(request.user)
     return render(request,'tarefas/listar_tarefas.html', {'tarefas': tarefas})
 
 
@@ -21,7 +23,7 @@ def cadastrar_tarefas(request):
             data_expiracao = form_tarefa.cleaned_data["data_expiracao"]
             prioridade = form_tarefa.cleaned_data["prioridade"]
             tarefanova = Tarefa(titulo = titulo, descricao = descricao, data_expiracao = data_expiracao,
-                                prioridade = prioridade)
+                                prioridade = prioridade, usuario = request.user)
             tarefa_service.cadastrar_tarefa(tarefanova)
             return redirect('listar_tarefas')
     else:
@@ -32,6 +34,8 @@ def cadastrar_tarefas(request):
 @login_required()
 def editar_tarefa(request, id):
     tarefa_bd = tarefa_service.listar_tarefa_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse("Não Permitido")
     form_tarefa = TarefaForm(request.POST or None, instance=tarefa_bd)
     if form_tarefa.is_valid():
         titulo = form_tarefa.cleaned_data["titulo"]
@@ -39,7 +43,7 @@ def editar_tarefa(request, id):
         data_expiracao = form_tarefa.cleaned_data["data_expiracao"]
         prioridade = form_tarefa.cleaned_data["prioridade"]
         tarefa_nova = Tarefa(titulo = titulo, descricao = descricao, data_expiracao = data_expiracao,
-                                prioridade = prioridade)
+                                prioridade = prioridade, usuario = request.user)
         tarefa_service.editar_tarefa(tarefa_bd, tarefa_nova)
         return redirect('listar_tarefas')
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
@@ -48,6 +52,9 @@ def editar_tarefa(request, id):
 @login_required()
 def remover_tarefa(request, id):
     tarefa_bd = tarefa_service.listar_tarefa_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse("Não Permitido")
+    form_tarefa = TarefaForm(request.POST or None, instance=tarefa_bd)
     if request.method == 'POST':
         tarefa_service.remover_tarefa(tarefa_bd)
         return redirect('listar_tarefas')
